test(stackable-versioned-macros): move tests without assertions into trybuild pass/fail examples
