+++ Cargo.toml
@@ -70,6 +70,7 @@ tracing = "0.1.40"
 tracing-appender = "0.2.3"
 tracing-opentelemetry = "0.24.0"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter", "json"] }
+trybuild = "1.0.99"
 url = { version = "2.5.2", features = ["serde"] }
 x509-cert = { version = "0.2.5", features = ["builder"] }
 zeroize = "1.8.1"
