+++ crates/stackable-versioned-macros/Cargo.toml
@@ -22,3 +22,4 @@ quote.workspace = true
 
 [dev-dependencies]
 rstest.workspace = true
+trybuild.workspace = true
