+++ crates/stackable-versioned-macros/tests/bad/README.md
@@ -1,0 +1,6 @@
+# trybuild bad cases
+
+Code that is expected to fail lives here along with the expected compiler output
+for each case. Please see the docs in [tests/trybuild.rs].
+
+[tests/trybuild.rs]: ../trybuild.rs
