+++ crates/stackable-versioned-macros/tests/bad/deprecate.stderr
@@ -1,0 +1,5 @@
+error: deprecation must be done using #[versioned(deprecated(since = "VERSION"))]
+  --> tests/bad/deprecate.rs:10:9
+   |
+10 |         #[deprecated]
+   |         ^
