+++ crates/stackable-versioned-macros/tests/bad/skip_from_version.stderr
@@ -1,0 +1,33 @@
+error[E0308]: mismatched types
+  --> tests/bad/skip_from_version.rs:23:32
+   |
+23 |     let foo_v1 = v1::Foo::from(foo_v1beta1);
+   |                  ------------- ^^^^^^^^^^^ expected `main::v1::Foo`, found `v1beta1::Foo`
+   |                  |
+   |                  arguments to this function are incorrect
+   |
+   = note: `v1beta1::Foo` and `main::v1::Foo` have similar names, but are actually distinct types
+note: `v1beta1::Foo` is defined in module `crate::main::v1beta1` of the current crate
+  --> tests/bad/skip_from_version.rs:4:5
+   |
+4  | /     #[versioned(
+5  | |         version(name = "v1alpha1"),
+6  | |         version(name = "v1beta1", skip(from)),
+7  | |         version(name = "v1")
+8  | |     )]
+   | |______^
+note: `main::v1::Foo` is defined in module `crate::main::v1` of the current crate
+  --> tests/bad/skip_from_version.rs:4:5
+   |
+4  | /     #[versioned(
+5  | |         version(name = "v1alpha1"),
+6  | |         version(name = "v1beta1", skip(from)),
+7  | |         version(name = "v1")
+8  | |     )]
+   | |______^
+note: associated function defined here
+  --> $RUST/core/src/convert/mod.rs
+   |
+   |     fn from(value: T) -> Self;
+   |        ^^^^
+   = note: this error originates in the attribute macro `versioned` (in Nightly builds, run with -Z macro-backtrace for more info)
