+++ crates/stackable-versioned-macros/tests/good/README.md
@@ -1,0 +1,6 @@
+# trybuild good cases
+
+Code that is expected to compile lives here. Please see the docs in
+[tests/trybuild.rs].
+
+[tests/trybuild.rs]: ../trybuild.rs
